requirements.txt
# Web Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1

# Database
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
alembic==1.13.1
GeoAlchemy2==0.14.3

# Data Validation and Settings
pydantic==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.1

# Geographical Data
shapely==2.0.3
geojson==3.1.0
mapbox==0.18.1

# Testing
pytest==8.0.0
httpx==0.26.0
pytest-asyncio==0.23.5

# Code Quality
black==24.1.1
flake8==7.0.0
isort==5.13.2

# API Documentation
python-multipart==0.0.9
python-jose[cryptography]==3.3.0

# Utilities
pandas==2.2.0
numpy==1.26.3
python-dateutil==2.8.2 